+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/State/Query.hs
@@ -13,16 +13,28 @@ module Cardano.Ledger.Api.State.Query (
   -- * Pool queries
   module Cardano.Ledger.Api.State.Query.Pool,
 
+  -- * Reward queries
+  module Cardano.Ledger.Api.State.Query.Reward,
+
   -- * @GetStakeSnapshots@
   module Cardano.Ledger.Api.State.Query.Snapshot,
 
   -- * @GetFilteredDelegationsAndRewardAccounts@
   module Cardano.Ledger.Api.State.Query.StakeDelegation,
+
+  -- * UTxO queries
+  module Cardano.Ledger.Api.State.Query.UTxO,
+
+  -- * Debug queries
+  module Cardano.Ledger.Api.State.Query.Debug,
 ) where
 
+import Cardano.Ledger.Api.State.Query.Debug
 import Cardano.Ledger.Api.State.Query.Epoch
 import Cardano.Ledger.Api.State.Query.Governance
 import Cardano.Ledger.Api.State.Query.PParams
 import Cardano.Ledger.Api.State.Query.Pool
+import Cardano.Ledger.Api.State.Query.Reward
 import Cardano.Ledger.Api.State.Query.Snapshot
 import Cardano.Ledger.Api.State.Query.StakeDelegation
+import Cardano.Ledger.Api.State.Query.UTxO